backend/requirements.txt
Django>=3.2,<4.0
djangorestframework>=3.12.4,<4.0
djangorestframework-simplejwt>=4.7.2,<5.0